Converting .format()'s to f-stings and removing RuntimeErrors
F-strings are easier to understand more consise and more efficient.

RuntimeErrors should not be used in our code. So I have created
a custom error type PSDSEvalError that is more appropriate as
all of the raised exceptions relate to poor or incorrect function
inputs.

I have also improved and added docstrings to the tests.

Test Plan:
I updated the tests and then ran them along with the example code
to make sure that all tests pass and all text printed was correct.

Change-Id: I369d3adcf39aaae6e1fd6bdc70307e410ef4cc83